simplified bms code
the previous approach was  managing recordd memory directly and sharing it across threads, and avoiding certain copies.

in order for too little gain and premature optimization was proving hard to debug in multithreaded environment, so it has been simplified overall.

currently deadlocks should be avoided and code should be optimal enough, no need to prematurely optimize