major refactoring of  the inference machine
replaced a lot of code to make multithreading more robust and less reliable on unsafe code, most important types are now behind an atomic rc but the code is still optimal as it's mostly handled through references.

major types are send + sync safe now and there should not be any UB happening.